build: Extract script utils to separate directory
