+++ scripts/utils/copy-assets.mts
