+++ scripts/utils/paths.mts
@@ -1,6 +1,6 @@
 import path from 'node:path';
 
-export const rootDir = path.join(import.meta.dirname, '..');
+export const rootDir = path.join(import.meta.dirname, '../..');
 
 export const buildDir = path.join(rootDir, 'build');
 export const genDir = path.join(rootDir, 'gen');
